fix: move path of supabase util file
