+++ src/pages/admin/content.tsx
@@ -14,7 +14,7 @@ import DeleteModal from '../../components/Dialogs/DeleteModal';
 import UpdateModal from '../../components/Dialogs/UpdateModal';
 import AdminLayout from '../../layouts/AdminLayout';
 import LinkModal from '../../components/Dialogs/LinkModal';
-import useSupabase from '../../hooks/useSupabase';
+import { useSupabase } from '../../utils/supabase';
 import { SchemaType } from '../../types/type-util';
 
 const Header = styled.div`
